refactor(web): kebab to PascalCase
